Site Editor: Merge the post only mode and the post editor
