+++ packages/edit-site/src/components/block-editor/site-editor-canvas.js
@@ -7,12 +7,9 @@ import classnames from 'classnames';
  */
 import { useSelect, useDispatch } from '@wordpress/data';
 import { useRef } from '@wordpress/element';
-import {
-	BlockList,
-	BlockTools,
-	store as blockEditorStore,
-} from '@wordpress/block-editor';
+import { BlockTools, store as blockEditorStore } from '@wordpress/block-editor';
 import { useViewportMatch, useResizeObserver } from '@wordpress/compose';
+
 /**
  * Internal dependencies
  */
@@ -28,12 +25,6 @@ import {
 } from '../../utils/constants';
 import { unlock } from '../../lock-unlock';
 import PageContentFocusNotifications from '../page-content-focus-notifications';
-
-const LAYOUT = {
-	type: 'default',
-	// At the root level of the site editor, no alignments should be allowed.
-	alignments: [],
-};
 
 export default function SiteEditorCanvas() {
 	const { clearSelectedBlock } = useDispatch( blockEditorStore );
@@ -56,16 +47,6 @@ export default function SiteEditorCanvas() {
 
 	const settings = useSiteEditorSettings();
 
-	const { hasBlocks } = useSelect( ( select ) => {
-		const { getBlockCount } = select( blockEditorStore );
-
-		const blocks = getBlockCount();
-
-		return {
-			hasBlocks: !! blocks,
-		};
-	}, [] );
-
 	const isMobileViewport = useViewportMatch( 'small', '<' );
 	const enableResizing =
 		isFocusMode &&
@@ -75,17 +56,7 @@ export default function SiteEditorCanvas() {
 
 	const contentRef = useRef();
 	const isTemplateTypeNavigation = templateType === NAVIGATION_POST_TYPE;
-
 	const isNavigationFocusMode = isTemplateTypeNavigation && isFocusMode;
-
-	// Hide the appender when:
-	// - In navigation focus mode (should only allow the root Nav block).
-	// - In view mode (i.e. not editing).
-	const showBlockAppender =
-		( isNavigationFocusMode && hasBlocks ) || isViewMode
-			? false
-			: undefined;
-
 	const forceFullHeight = isNavigationFocusMode;
 
 	return (
@@ -126,23 +97,6 @@ export default function SiteEditorCanvas() {
 									contentRef={ contentRef }
 								>
 									{ resizeObserver }
-									<BlockList
-										className={ classnames(
-											'edit-site-block-editor__block-list wp-site-blocks',
-											{
-												'is-navigation-block':
-													isTemplateTypeNavigation,
-											}
-										) }
-										dropZoneElement={
-											// Pass in the html element of the iframe to ensure that
-											// the drop zone extends to the very edges of the iframe,
-											// even if the template is shorter than the viewport.
-											contentRef.current?.parentNode
-										}
-										layout={ LAYOUT }
-										renderAppender={ showBlockAppender }
-									/>
 								</EditorCanvas>
 							</ResizableEditor>
 						</BlockTools>
