+++ packages/edit-site/src/components/block-editor/style.scss
@@ -14,7 +14,7 @@
 
 // Navigation focus mode requires padding around the root Navigation block
 // for presentational purposes.
-.edit-site-block-editor__block-list.is-navigation-block {
+.edit-site-editor-canvas__block-list.is-navigation-block {
 	padding: $grid-unit-30;
 }
