+++ packages/editor/src/store/private-actions.js
@@ -1,0 +1,13 @@
+/**
+ * Returns an action object used to set which template is currently being used/edited.
+ *
+ * @param {string} id Template Id.
+ *
+ * @return {Object} Action object.
+ */
+export function setCurrentTemplateId( id ) {
+	return {
+		type: 'SET_CURRENT_TEMPLATE_ID',
+		id,
+	};
+}
