+++ packages/editor/src/store/reducer.js
@@ -90,6 +90,15 @@ export function postId( state = null, action ) {
 	return state;
 }
 
+export function templateId( state = null, action ) {
+	switch ( action.type ) {
+		case 'SET_CURRENT_TEMPLATE_ID':
+			return action.id;
+	}
+
+	return state;
+}
+
 export function postType( state = null, action ) {
 	switch ( action.type ) {
 		case 'SETUP_EDITOR_STATE':
@@ -291,6 +300,7 @@ export function renderingMode( state = 'all', action ) {
 export default combineReducers( {
 	postId,
 	postType,
+	templateId,
 	saving,
 	deleting,
 	postLock,
