+++ packages/editor/src/store/selectors.js
@@ -206,6 +206,17 @@ export function getCurrentPostId( state ) {
 }
 
 /**
+ * Returns the template ID currently being rendered/edited
+ *
+ * @param {Object} state Global application state.
+ *
+ * @return {string?} Template ID.
+ */
+export function getCurrentTemplateId( state ) {
+	return state.templateId;
+}
+
+/**
  * Returns the number of revisions of the post currently being edited.
  *
  * @param {Object} state Global application state.
